Extract permission related code to new file
